fixing suspense
